Refactor NAPALM get_bgp_neighbors(_detail) methods
The current _detail implementation produces invalid output, as follows:
* When 2 peerings are present (ipv4/ipv6), only 1 is reported
* Name of the default VRF should be 'global', not 'Base'
* local_address missing
* local_address_configured wrong ( True when it should be False )
* router_id should be the peer's router id, not the local router_id
* multipath is true unless explicitly disabled
* suppress_4byte_as is false unless explicitly configured otherwise
* local_port is retrieved from the wrong context (value from vprn presented in Base)
* remote_address is set to peer router_id instead of interface address
* configured_keepalive is set to operational keepalive
* keepalive is missing
* holdtime is missing
* accepted_prefix_count not set (PR uses received - rejected count)
* routing_table is wrong (should be name of the VRF, not peering group)
* suppress_4byte_as is wrong

The get_bgp_neighbors method behaves similar to before, except that the 'management' vrf is no longer included
in the output. Its implementation is aligned with the _detail method approach

Given the extent of deviations, this PR provides refactored methods that are placed in separate files.
This provides improved maintainability, as the methods can be tested in isolation and are self-contained

It also fixes a recent change in the 'speed' data type which should now be float, not int